fix: resolve type mismatch errors (76→65)
Fixed multiple E0308 type mismatch errors:
- syscall/mod.rs: Changed table.get(&pid) → table.get(pid) (4 instances)
- procfs.rs: Fixed table.get(&pid) and table.get(&self.pid) (6 instances)
- mm/mod.rs: Cast num_pages to u8 for alloc_pages()
- mm/mod.rs: Cast KERNEL_BASE to usize in phys_to_virt()
- ai_dashboard.rs: Convert string literals to String for Label::new()

Progress: 238 → 65 errors (173 fixed, 73% complete)

+++ crates/kernel/src/mm/mod.rs
@@ -62,7 +62,7 @@ pub use aslr::{
 ///
 /// Used by DMA devices that need contiguous physical memory
 pub fn alloc_phys_pages(num_pages: usize) -> Option<PhysAddr> {
-    let page = alloc_pages(num_pages)?;
+    let page = alloc_pages(num_pages as u8)?;
     Some(page as u64)
 }
 
@@ -70,5 +70,5 @@ pub fn alloc_phys_pages(num_pages: usize) -> Option<PhysAddr> {
 ///
 /// Assumes direct mapping: phys_addr + KERNEL_BASE = virt_addr
 pub fn phys_to_virt(phys: PhysAddr) -> usize {
-    (phys as usize).wrapping_add(KERNEL_BASE)
+    (phys as usize).wrapping_add(KERNEL_BASE as usize)
 }
